+++ packages/oauth/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.3.0
+
+- [Breaking] Rename type `GetUserType` to `LuciaUser`; remove `GetCreateUserAttributesType`
+- Make `Buffer` dependency optional
+
 ## 0.2.7
 
 - Fix type issues with `existingUser` and `createUser()` for `validateCallback()`